Enum class

+++ src/Test/main.cpp
@@ -16,7 +16,7 @@ void printMachineState(Logger &logger, std::shared_ptr<Machine> machine) {
 }
 
 void printActionResults(Logger &logger, ActionStatusMessage &message) {
-    logger.Log("[ACM] " + std::to_string(message.getType()) + " | " + message.getContent());
+    logger.Log("[ACM] " + std::to_string(static_cast<double>(message.getType())) + " | " + message.getContent());
 }
 
 class CountListener : public CountMessageReceiver {
@@ -27,7 +27,8 @@ public:
     CountListener(const Logger &logger, TempoComponent &component) : m_logger(logger), m_component(component) {}
 
     void ReceiveMessage(std::shared_ptr<CountMessage> ptr) override {
-        m_logger.Log("[CL] Component " + m_component.getName() + " says : " + std::to_string(ptr->getCount()));
+        m_logger.Log("[CL] Component " +
+        m_component.getName() + " says : " + std::to_string(ptr->getCount()));
     }
 };
 
@@ -37,28 +38,34 @@ int main() {
         std::shared_ptr<Machine> machine = std::make_shared<Machine>("Test Machine 1");
         printMachineState(logger, machine);
 
-        std::shared_ptr<Feeder> feeder = std::make_shared<Feeder>("Feeder Test 1", *machine.get(), 17000, 100);
+        std::shared_ptr<Feeder> feeder = std::make_shared<Feeder>("Feeder Test 1", *machine.get(),
+                17000, 100);
         machine->addComponent(feeder);
         std::shared_ptr<CountMessageReceiver> listener = std::make_shared<CountListener>(logger, *feeder.get());
         feeder->getCountMessageReceiver().push_back(listener);
 
-        std::shared_ptr<PaintStation> cyan = std::make_shared<PaintStation>("PaintStation Cyan Test 1", *machine.get(), 10000, 10000);
+        std::shared_ptr<PaintStation> cyan = std::make_shared<PaintStation>("PaintStation Cyan Test 1",
+                *machine.get(), 10000, 10000);
         machine->addComponent(cyan);
 
-        std::shared_ptr<PaintStation> magenta = std::make_shared<PaintStation>("PaintStation Magenta Test 1", *machine.get(), 10000, 10000);
+        std::shared_ptr<PaintStation> magenta = std::make_shared<PaintStation>("PaintStation Magenta Test 1",
+                *machine.get(), 10000, 10000);
         machine->addComponent(magenta);
 
-        std::shared_ptr<PaintStation> yellow = std::make_shared<PaintStation>("PaintStation Yellow Test 1", *machine.get(), 10000, 10000);
+        std::shared_ptr<PaintStation> yellow = std::make_shared<PaintStation>("PaintStation Yellow Test 1",
+                *machine.get(), 10000, 10000);
         machine->addComponent(yellow);
 
-        std::shared_ptr<PaintStation> black = std::make_shared<PaintStation>("PaintStation Black Test 1", *machine.get(), 10000, 10000);
+        std::shared_ptr<PaintStation> black = std::make_shared<PaintStation>("PaintStation Black Test 1",
+                *machine.get(), 10000, 10000);
         machine->addComponent(black);
 
-        std::shared_ptr<Delivery> delivery = std::make_shared<Delivery>("Delivery Test 1", *machine.get(), 17000, 16950);
+        std::shared_ptr<Delivery> delivery = std::make_shared<Delivery>("Delivery Test 1",
+                *machine.get(), 17000, 16950);
         machine->addComponent(delivery);
 
-        std::shared_ptr<Conveyor> conveyor = std::make_shared<Conveyor>("Conveyor Test 1", *machine.get(), 15000,
-                                                                        15000);
+        std::shared_ptr<Conveyor> conveyor = std::make_shared<Conveyor>("Conveyor Test 1",
+                *machine.get(), 15000, 15000);
         machine->addComponent(conveyor);
 
         printMachineState(logger, machine);
